backend/requirements-macos.txt
# Python 3.9.6
absl-py==1.0.0
appnope==0.1.3
asttokens==2.0.5
astunparse==1.6.3
autopep8==1.7.0
backcall==0.2.0
CacheControl==0.12.11
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==38.0.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
firebase-admin==6.0.1
flatbuffers==2.0
fonttools==4.37.1
gast==0.5.3
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
grpcio==1.50.0
h5py==3.7.0
httplib2==0.20.4
idna==3.3
imageio==2.35.1
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jupyter-client==7.3.1
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.4
libclang==14.0.1
load-dotenv==0.1.0
Markdown==3.3.7
matplotlib==3.5.3
matplotlib-inline==0.1.3
msgpack==1.0.4
nest-asyncio==1.5.5
networkx==3.2.1
numpy==1.23.4
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.29
proto-plus==1.22.1
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
pyzmq==23.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.24.0
scipy==1.13.1
Shapely==1.8.2
six==1.15.0
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-macos==2.8.0
tensorflow-metal==0.4.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2024.8.30
toml==0.10.2
tornado==6.1
traitlets==5.2.1.post0
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
